requirements.txt
matplotlib==3.7.1
wordcloud==1.9.2
jieba==0.42.1
numpy==1.24.3
pillow==9.5.0
pandas==2.0.1
seaborn==0.12.2
requests==2.31.0
beautifulsoup4==4.12.2